runner: allow request() to be external
Instead of injecting (optionally) a control channel, inject a request
method. This works better, because a method can easily be provided, but
the actualy control channel with the new WS API changes with each new
supervisor that connects.

Longer term, any dependence on the control channel should be removed
from the runner, which can just run apps. This is easy, except the tests
currently depend heavily on messaging with the supervisor, they'd have
to be written to create a WS router and receive control channels, which
is too time consuming for the moment.